requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.30
botocore==1.37.30
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-aws==0.2.18
langchain-core==0.3.51
langchain-mcp-adapters==0.0.7
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langgraph==0.3.27
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.27
mcp==1.6.0
numpy==2.2.4
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.1
tenacity==9.1.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uv==0.6.13
uvicorn==0.34.0
xxhash==3.5.0
zstandard==0.23.0
